Cut up tests into seperate classes
